fix: dashboard page detail pane extract to components and composables
