+++ packages/web/design-system/src/ui/dashboard-page/DashboardPage.vue
@@ -3,6 +3,7 @@ import { useTitle } from '@vueuse/core'
 import {
   computed,
   onUnmounted,
+  ref,
   useSlots,
   watch,
 } from 'vue'
@@ -55,6 +56,8 @@ watch([
   immediate: true,
 })
 
+const isDetailPaneOpen = ref<boolean>(true)
+
 const hasDetailPane = computed<boolean>(() => {
   return props.detailPane !== null && slots['detail-pane'] !== undefined
 })
@@ -70,6 +73,7 @@ const {
   onResizeKeyDown,
   onResizeStart,
 } = useDetailPane({
+  isOpen: isDetailPaneOpen,
   isResizable: props.detailPane?.isResizable ?? true,
   storage: props.detailPane?.storage ?? null,
   variant: props.detailPane?.variant ?? 'full-height-inline',
