+++ packages/web/design-system/src/ui/dashboard-page/centered-content/DashboardPageCenteredContentHeaderScrollGradient.vue
@@ -1,0 +1,40 @@
+<script setup lang="ts">
+
+</script>
+
+<template>
+  <div
+    aria-hidden="true"
+    class="pointer-events-none absolute inset-0"
+  >
+    <div class="absolute inset-0 bg-linear-to-b from-primary to-transparent" />
+    <div
+      class="
+        absolute inset-0
+        mask-[linear-gradient(to_bottom,black_0%,transparent_100%)]
+        backdrop-blur-xs
+      "
+    />
+    <div
+      class="
+        absolute inset-0
+        mask-[linear-gradient(to_bottom,black_0%,transparent_75%)]
+        backdrop-blur-xs
+      "
+    />
+    <div
+      class="
+        absolute inset-0
+        mask-[linear-gradient(to_bottom,black_0%,transparent_50%)]
+        backdrop-blur-sm
+      "
+    />
+    <div
+      class="
+        absolute inset-0
+        mask-[linear-gradient(to_bottom,black_0%,transparent_25%)]
+        backdrop-blur-md
+      "
+    />
+  </div>
+</template>
